fix: enforce RBAC, decimal quantities, and permission scoping
  Permission fixes:
  - Remove GROUP_LEAD from global bin transaction gate; scope checkout to
    self-service, return to owner/same-group, list to own/group/all by role
  - Restrict usage event creation to can_manage_inventory (block GROUP_LEAD)
  - Require actual checklist assignment (not just group_name) to complete tasks
  - Block GROUP_LEAD from assigning/managing across group boundaries in
    all checklist management paths (_can_manage_tasks_on, assign, unassign,
    add_checklist_item)
  - Fix return_transaction to pass borrower's group so GROUP_LEAD can return
    same-group item transactions (consistent with checkout)

  Decimal quantity support:
  - Convert purchase_records.quantity_purchased to Numeric(10,2)
  - Update purchase schema (Decimal create, float out) and fix float*Decimal
    arithmetic in purchase_service and reports router
  - Change report schemas (total_purchased, total_used, quantity fields) to float
  - Add migration 017 for purchase_records.quantity_purchased column change

  Other:
  - Decimal quantities in usage, stock adjustment, inventory request schemas
  - ActivityLog.quantity_delta changed to Numeric(10,2) (migration 015)
  - Migration 016 removes global powers from GROUP_LEAD role
  - Add 17 focused regression tests covering all fixed permission paths
  - Rooms support, checklists subchecklists, checklist weekly reset (Chicago TZ)
  - Telegram HTML escaping, snake_case checklist API params, ESLint config

+++ backend/app/models/activity_log.py
@@ -73,7 +73,7 @@ class ActivityLog(Base):
     target_user_id: Mapped[Optional[int]] = mapped_column(ForeignKey("users.id"), nullable=True)
 
     # Quantity / cost context (signed: positive = increase, negative = decrease)
-    quantity_delta: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
+    quantity_delta: Mapped[Optional[float]] = mapped_column(Numeric(10, 2), nullable=True)
     cost_impact: Mapped[Optional[float]] = mapped_column(Numeric(10, 2), nullable=True)
 
     notes: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
