+++ backend/app/models/transaction.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from enum import Enum
 from typing import TYPE_CHECKING, List, Optional
 
-from sqlalchemy import Boolean, CheckConstraint, DateTime, ForeignKey, Integer, String, Text, func
+from sqlalchemy import Boolean, CheckConstraint, DateTime, ForeignKey, Integer, Numeric, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -56,7 +56,7 @@ class Transaction(Base):
         ForeignKey("bin_transactions.id"), nullable=True, index=True
     )
 
-    quantity: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
+    quantity: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False, default=1)
     status: Mapped[str] = mapped_column(
         String(20), nullable=False, default=TransactionStatus.CHECKED_OUT, index=True
     )
