+++ backend/app/schemas/stock_adjustment.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel, ConfigDict, Field
@@ -6,7 +7,7 @@ from pydantic import BaseModel, ConfigDict, Field
 
 class StockAdjustmentCreate(BaseModel):
     item_id: int
-    delta: int = Field(..., description="Signed quantity change: positive=add, negative=remove")
+    delta: Decimal = Field(..., description="Signed quantity change: positive=add, negative=remove", decimal_places=2)
     reason: str = "CORRECTION"
     notes: Optional[str] = None
 
@@ -17,9 +18,9 @@ class StockAdjustmentOut(BaseModel):
     id: int
     item_id: int
     adjusted_by_user_id: int
-    delta: int
-    quantity_before: int
-    quantity_after: int
+    delta: Decimal
+    quantity_before: Decimal
+    quantity_after: Decimal
     reason: str
     notes: Optional[str] = None
     adjusted_at: datetime
